+++ app/src/main/java/com/example/openweatherapi/data/remote/dto/ListItem.kt
@@ -7,13 +7,13 @@ import com.google.gson.annotations.SerializedName
 
 
 data class ListItem(
-    val dt: Long,
+    val dt: Int,
     val main: Main2,
     val weather: ArrayList<Weather> = arrayListOf(),
     val clouds: Clouds,
     val wind: Wind,
     val visibility: Int,
-    val pop: Int,
+    val pop: Double,
     val sys: Sys2,
     @SerializedName("dt_txt")
     val dtTxt: String
